+++ config/eslintrc.js
@@ -129,6 +129,15 @@ module.exports = {
             'error',
             { allow: ['arrowFunctions', 'constructors'] },
         ],
+        'no-return-await': 'off',
+        '@typescript-eslint/no-unused-vars': [
+            'error',
+            {
+                args: 'all',
+                argsIgnorePattern: '^_',
+                varsIgnorePattern: '^_',
+            },
+        ],
     },
     overrides: [
         {