+++ Changelog.md
@@ -7,6 +7,14 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## Unreleased
+
+### Changed
+
+- Disabled deprecated `no-return-await` ESLint rule.
+- Changed `no-unused-vars` ESLint rule config so that only unused args and vars
+  beginning with an underscore are allowed.
+
 ## 231.0.0 - 2025-10-07
 
 ### Changed